crud_sqlite-master/requirements.txt
Flask==3.0.3
flask_sqlalchemy==3.1.1
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-Limiter==3.7.0
python-dotenv==1.0.1
pytest==8.3.2
